# public/openapi/investor-match-schema.yaml
openapi: 3.0.0
info:
  title: Investor Matching API
  description: An API to match real estate deals with potential investors based on preferences and criteria.
  version: 1.0.0
servers:
  - url: https://capital-match-ai-platform.vercel.app/api
    description: Production server
paths:
  /investor-match:
    post:
      summary: Find investors that match a given real estate deal
      operationId: matchInvestors
      tags:
        - Investor Matching
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - assetType
                - market
                - investmentAmount
                - expectedReturn
                - riskProfile
              properties:
                assetType:
                  type: string
                  description: The type of real estate asset
                  example: "multifamily"
                  enum:
                    - multifamily
                    - office
                    - industrial
                    - retail
                    - hotel
                    - student housing
                    - mixed-use
                market:
                  type: string
                  description: The geographic market/city where the property is located
                  example: "Chicago"
                investmentAmount:
                  type: number
                  description: The amount of investment required in USD
                  example: 5000000
                expectedReturn:
                  type: number
                  description: The expected annual return percentage
                  example: 12
                riskProfile:
                  type: string
                  description: The risk level of the investment
                  example: "moderate"
                  enum:
                    - low
                    - moderate
                    - high
                projectName:
                  type: string
                  description: Optional name of the project/property
                  example: "Lakeview Apartments"
                description:
                  type: string
                  description: Optional brief description of the investment opportunity
                  example: "256-unit apartment complex in Lakeview neighborhood"
                timeline:
                  type: string
                  description: Optional timeline for the investment
                  example: "5 years"
      responses:
        '200':
          description: Successfully matched investors with the deal
          content:
            application/json:
              schema:
                type: object
                properties:
                  deal:
                    type: object
                    properties:
                      assetType:
                        type: string
                        example: "multifamily"
                      market:
                        type: string
                        example: "Chicago"
                      investmentAmount:
                        type: number
                        example: 5000000
                      expectedReturn:
                        type: number
                        example: 12
                      riskProfile:
                        type: string
                        example: "moderate"
                  matches:
                    type: array
                    items:
                      type: object
                      properties:
                        investorId:
                          type: string
                          example: "INV001"
                        investorName:
                          type: string
                          example: "Blackstone Real Estate"
                        matchScore:
                          type: number
                          example: 85
                        matchDetails:
                          type: object
                          properties:
                            assetTypeMatch:
                              type: boolean
                              example: true
                            marketMatch:
                              type: boolean
                              example: true
                            investmentSizeMatch:
                              type: boolean
                              example: true
                            returnExpectationMatch:
                              type: boolean
                              example: false
                            riskProfileMatch:
                              type: boolean
                              example: true
                  totalMatches:
                    type: number
                    example: 3
        '400':
          description: Bad request - missing required parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Missing required field: assetType"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "An error occurred while processing your request"